+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Improve the plugin dev support docs and add a pluginConfigMutator option
+- Add a whitelist mechanism to control when the sidebar appears in farmOS
 
 ### Changed